Refactored main file
