+++ .gitignore
@@ -134,3 +134,6 @@ hclasses/testhclasses
 core-adagucserver*
 data/datasets/arcus_multivar/uwcw_ha43_dini_testfile_multivar_5x6_clean.nc
 tests/env_checker_report.txt
+pgbouncer.log
+pgbouncer.pid
+info.c
