+++ adagucserverEC/geojsondump.cpp
@@ -5,7 +5,6 @@
 #include "CDebugger.h"
 #include "CCDFDataModel.h"
 #include "CCDFGeoJSONIO.h"
-DEF_ERRORMAIN();
 
 int main(int argCount, char **argVars) {
   CDFGeoJSONReader *geoJSONReader = NULL;
