+++ adagucserverEC/utils/processCMDArgs.h
@@ -1,0 +1,1 @@
+int processCMDArgs(int argc, char **argv, char **);
