+++ hclasses/CDebugger.h
@@ -78,6 +78,5 @@ void _printError(const char *pszMessage, ...);
   const char *functionName = name;                                                                                                                                                                     \
   _printDebugLine("D %s, %d class %s: Entering function '%s'", __FILENAME__, __LINE__, className, functionName);
 #define DEF_ERRORFUNCTION() static const char *className;
-#define DEF_ERRORMAIN()
 
 #endif
